+++ Text Formatting/Banks/CheckAccTF.h
@@ -8,5 +8,5 @@ using namespace std;
 
 class CheckAccTF {
 public:
-	void BankLog(string name, int amount);
+	void BankLog(string name, float amount);
 };
